+++ src/storage-blob-preview/azext_storage_blob_preview/_params.py
@@ -494,7 +494,13 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('data', help='The blob data to upload.', required=False, is_preview=True, min_api='2019-02-02')
         c.argument('length', type=int, help='Number of bytes to read from the stream. This is optional, but should be '
                    'supplied for optimal performance. Cooperate with --data.', is_preview=True, min_api='2019-02-02')
-        c.argument('overwrite', overwrite_type, arg_group="Additional Flags", is_preview=True)
+        c.argument('overwrite', arg_type=get_three_state_flag(), arg_group="Additional Flags", is_preview=True,
+                   help='Whether the blob to be uploaded should overwrite the current data. If True, upload_blob will '
+                   'overwrite the existing data. If set to False, the operation will fail with ResourceExistsError. '
+                   'The exception to the above is with Append blob types: if set to False and the data already exists, '
+                   'an error will not be raised and the data will be appended to the existing blob. If set '
+                   'overwrite=True, then the existing append blob will be deleted, and a new one created. '
+                   'Defaults to False.')
         c.argument('max_connections', type=int, arg_group="Additional Flags",
                    help='Maximum number of parallel connections to use when the blob size exceeds 64MB.')
         c.extra('maxsize_condition', type=int, arg_group="Content Control",